System.Process

Description

This library provides process spawning facilities for POSIX systems.

Features

Always reaps children spawned by this library, no need to explicitly wait for each child if you don't care about the result.

Uses ByteStrings for OS strings.

Works with the non-threaded and threaded runtimes.

Uses SIGCHILD to detect child process termination.

Can connect pipes at specifc file descriptors.

Can connect pipes to free file descriptors and communicate the file descriptor to the child process using environment variables.

Signals are always sent to the intended process, even if your system reuses PIDs.

Types help you to ensure you feed and read all pipes.

Pipes are guaranteed to only go to their destined process, regardless of spawning processes in parallel (as long as you only fork processes with this library).

Currently it only works on POSIX systems.

NOTE: This library modifies your SIGCHILD signal handler. It will not work if you reset this.

NOTE: If you use the fork() system call, we do not guarantee that file descriptors created by this library won't leak. In fact, we know of no way to guarantee this in a threaded program across all systems. If the only fork()ing you do is from this library, then we do ensure that file descriptors don't leak.

NOTE: If you create file descriptors in other parts of your program, you should bracket their creation inside the dontFork function and make sure that you have set the CLOEXEC flag before they return to prevent them leaking into process forked by this library.

Examples

Simple process execution, feeding stdin, and reading stdout.

>>> captureInOut (proc "md5sum" []) "Hello"
("8b1a9953c4611296a827abf8c47804d7  -\n",Exited ExitSuccess)

>>> captureInOutErr (proc "cat" ["-", "nofile"]) "Hello, world!"
(StdOutputs {getStdOut = "Hello, world!", getStdErr = "cat: nofile: No such file or directory\n"},Exited (ExitFailure 1))

Documentation

data ProcessSpec ipipes opipes

This type represents the description of a process that is going to be executed.

Constructors

ProcessSpec

Fields

command :: !ByteString

The command to run

arguments :: ![ByteString]

The positional arguments

searchPath :: !Bool

Should we search your PATH environment variable?

inputs :: !(ipipes PipeLocation)

The input pipes to create. The child will read these.

outputs :: !(opipes PipeLocation)

The output pipes to create. The child will write these.

data ProcessHandle

This type represents a running process. Use this to send signals, or wait for the process to finish.

See processWait, processTryWait and processSignal.

data PipeLocation

Defines where to create a pipe.

For some pre-constructed values of this, see stdOut, stdErr, stdIn and stdOutputs

Constructors

Dup2 !Fd
Move the pipe to the specified file descriptor.

Env !ByteString
Find a free file descriptor, and launch the process with the given environment variable set to the file descriptor number.

proc :: ByteString -> [ByteString] -> ProcessSpec NoPipe NoPipe

Create a ProcessSpec with no pipes. These can later be extended with various functions in this library. e.g. readPipe, writePipe, etc.

procWith :: i PipeLocation -> o PipeLocation -> ByteString -> [ByteString] -> ProcessSpec i o

Create a ProcessSpec with the given pipes.

>>> capture (procWith NoPipe stdOutputs "echo" ["Hello, world!"]) NoPipe
(StdOuputs { getStdOut = "Hello, world!\n", getStdErr = ""},Exited ExitSuccess)

These can still be extended with the readPipe, writePipe, etc. functions.

Adding pipes

These take a ProcessSpec and add additional pipes to it. You can instead use procWith to create all the pipes you want at once.

readPipe :: PipeLocation -> ProcessSpec i o -> ProcessSpec (AddPipe i) o

Create a pipe that the child can read from.

readPipes :: i PipeLocation -> ProcessSpec i' o -> ProcessSpec (Product i' i) o

Create potentially multiple pipes that the child can read from.

writePipe :: PipeLocation -> ProcessSpec i o -> ProcessSpec i (AddPipe o)

Create a pipe that the child can write to.

writePipes :: o PipeLocation -> ProcessSpec i o' -> ProcessSpec i (Product o' o)

Create potentially multiple pipes that the child can write to.

Running a process

spawn :: (Traversable i, Traversable o) => ProcessSpec i o -> IO (ProcessHandle, i Fd, o Fd)

This is the most primite spawning operation in this library. It takes a ProcessSpec and turns it into a process. You can interact with the process using the returned ProcessHandle and input and output file descriptors.

There is no need to call processWait unless you want to know the exit code and reason of the process. It will be reaped regardless. You probably do want to call closeFd on the file descriptors though.

capture :: (Pipes i, Pipes o) => ProcessSpec i o -> i ByteString -> IO (o ByteString, ProcessStatus)

A function that allows you to capture all the read pipes, and write to all the write pipes.

>>> capture (readPipe stdIn $ writePipe stdOut $ proc "xxd" []) "Hello" processWait
(APipe NoPipe "00000000: 4865 6c6c 6f                             Hello\n",Exited ExitSuccess)

captureIn :: ProcessSpec NoPipe NoPipe -> ByteString -> IO ProcessStatus

Spawn a process, and write to stdin.

captureOut :: ProcessSpec NoPipe NoPipe -> IO (ByteString, ProcessStatus)

Spawn a process and read from stdout.

captureInOut :: ProcessSpec NoPipe NoPipe -> ByteString -> IO (ByteString, ProcessStatus)

Spawn a process and write to stdin and read from stdout.

captureInOutErr :: ProcessSpec NoPipe NoPipe -> ByteString -> IO (StdOutputs ByteString, ProcessStatus)

Spawn a process and write to stdin, and read from both stdout and stderr.

handlePipes :: (Pipes i, Pipes o) => ProcessSpec i o -> i (Handle -> IO ai) -> o (Handle -> IO ao) -> (ProcessHandle -> IO ap) -> IO (i ai, o ao, ap)

Spawn the process, and run an IO action on each handle created. Note that this does not wait for the process to exit unless the one of the handlers does so.

handleIn :: ProcessSpec NoPipe NoPipe -> (Handle -> IO a) -> IO (a, ProcessStatus)

Run an IO with access to a Handle connected to stdin.

handleOut :: ProcessSpec NoPipe NoPipe -> (Handle -> IO a) -> IO (a, ProcessStatus)

Run an IO with access to a Handle connected to stdout.

handleInOut :: ProcessSpec NoPipe NoPipe -> (Handle -> IO a) -> (Handle -> IO b) -> IO (a, b, ProcessStatus)

Run two IO actions concurrently each with a Handle, one connected to stdin, the other to stdout.

Interacting with processes

processWait :: ProcessHandle -> IO ProcessStatus

Wait for this process to exit, and return it's exit status.

You do NOT have to call this in order to clean up your process. It will be reaped regardless.

Note: this call will not return until the process has exited, but it does NOT block the thread, and is safe to use in the single threaded runtime.

processTryWait :: ProcessHandle -> IO (Maybe ProcessStatus)

Like processWait, but does not block.

processSignal :: ProcessHandle -> Signal -> IO ()

Send a Signal to this process. The signal is guaranteed to be sent to exactly this process even if your system reuses process ID's.

Standard locations

Use these when constructing your pipes if you need to refer to standard locations.

stdOut :: PipeLocation

Creates a pipe connected to the stdout of the child process.

stdErr :: PipeLocation

Creates a pipe connected to the stderr of the child process.

stdIn :: PipeLocation

Creates a pipe connected to the stdin of the child process.

stdOutputs :: StdOutputs PipeLocation

A precomposed value of StdOutputs which refers to stdout and stderr and can be used directly when you want to capture those.

>>> capture (procWith NoPipe stdOutputs "echo" ["yes"]) NoPipe
(StdOuputs { getStdOut = "yes\n", getStdErr = ""},Exited ExitSuccess)

data StdOutputs a

Use stdOutputs to construct this. Will create pipes for stderr and stdout.

Constructors

StdOutputs

Fields

getStdOut :: a

getStdErr :: a

Composable types for adding pipes.

class Traversable a => Pipes a where

A class for containers that contain pipe locations.

This essentially just provides an aligned, uniform shaped zip function which is used to align the provided pipe handlers with the created pipes.

Methods

zipPipesM :: Monad m => (x -> y -> m z) -> a x -> a y -> m (a z)

data NoPipe a

A type to use when you don't want any pipes. Also used as the "Nil" element when you add pipes using AddPipe.

The proc command will create a ProcessSpec NoPipe NoPipe to signal that by default there are no pipes created. You can then use some of the combinators in this library to add new pipes, or you can bypass the proc call and use procWith instead to directly specify the pipes you want to create.

Constructors

NoPipe

data b :| a

Add another pipe onto a process. (See also the AddPipe alias which is easier to read in some scenarios.)

Constructors

(b a) :| a

type AddPipe = (:|)

Sometimes this is easier to read.

Miscellaneous

dontFork :: IO a -> IO a

This will run the given IO action while preventing this library from doing almost everything. It won't fork, it won't create pipes, and it won't reap children processes. IO actions should therefore remain very short. The purpose of this is for when you need to create a file descriptor that can not atomically have the CLOEXEC bit set, and so you need to do it non-atomically across multiple system calls.

The IO action MUST NOT call any other function from this library.